[skip ci] PHRAS-3430_create-ps-conf-tables_MASTER refacto with a "App" base class for app instances WIP
